Move RNSScreen enum type defs to external header
This allows to share those types between old and new architecture
